Moved hooks to internal directory
