+++ packages/component/src/hooks/internal/useForceRenderAtInterval.js
