+++ packages/component/src/hooks/internal/useTimer.js
